backend/requirements.txt
flask
flask-cors
textblob
gTTS
SpeechRecognition
pydub
torch
transformers
happytransformer
sentencepiece
metaphone
fuzzywuzzy
python-Levenshtein
pymupdf
requests
readability-lxml    # Extracting "Main Content" from web pages
beautifulsoup4      # For cleaning HTML tags
pymupdf             # (fitz) For extracting structure from PDFs
html2text           
nltk
